refactor(studio): split settings.rs and prefs.rs under the 1000-line limit
Addresses the Low review finding on PR #71. Both files were already over
the threshold before this PR; the provider-wiring additions pushed the
duplication into view. Pure code moves, no behaviour change:

- settings.rs (1069 → 808): provider/model resolution `impl Settings`
  block + preview-model fns → settings/providers.rs; thinking-budget /
  style combobox helpers → settings/model_defaults.rs. Re-exported so
  `crate::settings::*` callers are untouched.
- prefs.rs (1199 → 462): the ~700-line `prefs_tab_llm` method and its
  `price_grid_row` helper → app/ui_dialogs/prefs/llm.rs.
